Move MUX code to a separate file
The main file is already large; and we're going to develop the MUX emulation
further. Let's move all the relevant code to a separate file
