+++ Makefile
@@ -2,11 +2,13 @@ all: centurion
 
 CFLAGS = -g3 -Wall -pedantic
 
-centurion: centurion.o cpu6.o
+centurion: centurion.o cpu6.o mux.o
 
-centurion.o: centurion.c cpu6.h
+centurion.o: centurion.c centurion.h cpu6.h
 
 cpu6.o : cpu6.c cpu6.h
+
+mux.o : centurion.h mux.h
 
 clean:
 	rm -f centurion *.o *~
