+++ centurion.h
@@ -1,0 +1,1 @@
+extern volatile unsigned int emulator_done;
